Added enemy collision with player
Enemy now changes direction when collides with player. Player jumps to the middle of the screen and takes 10 damage from health. Also updated player movement mechanics so they move towards touch point gradually as opposed to jumping directly to touch point.

+++ core/src/com/becksm64/coingetter/Enemy.java
@@ -59,7 +59,7 @@ public class Enemy {
         return velocity;
     }
 
-    public void setVelocity(int velX, int velY) {
+    public void setVelocity(float velX, float velY) {
         this.velocity = new Vector3(velX, velY, 0);
     }
